requirements.txt
contourpy==1.3.0
cycler==0.12.1
eventkit==1.0.3
fonttools==4.54.1
ib-insync==0.9.86
kiwisolver==1.4.7
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.14.1
six==1.16.0
statsmodels==0.14.4
tzdata==2024.2
